external camel consumers
 - Removed dublicated dependencies
 - Added delayed startup of the containers to ease the load on the system when bringing up the containers
 - Fixed DefaultJaxbContextProvider configuration

+++ qa/common/src/main/java/org/eclipse/kapua/qa/common/DBHelper.java
@@ -153,15 +153,18 @@ public class DBHelper {
 
     public void dropAll() throws SQLException {
 
-        String[] types = {"TABLE"};
-        ResultSet sqlResults = connection.getMetaData().getTables(null, null, "%", types);
+        if (!connection.isClosed()) {
+            String[] types = {"TABLE"};
 
-        while (sqlResults.next()) {
-            String sqlStatement = String.format("DROP TABLE %s", sqlResults.getString("TABLE_NAME").toUpperCase());
-            connection.prepareStatement(sqlStatement).execute();
+            ResultSet sqlResults = connection.getMetaData().getTables(null, null, "%", types);
+
+            while (sqlResults.next()) {
+                String sqlStatement = String.format("DROP TABLE %s", sqlResults.getString("TABLE_NAME").toUpperCase());
+                connection.prepareStatement(sqlStatement).execute();
+            }
+
+            this.close();
         }
-
-        this.close();
         KapuaCacheManager.invalidateAll();
     }
 }
